Refactoring - rename PlanItem concept to PlanElement (Item and SubItem) in Student mode
