+++ src/screens/index.tsx
@@ -8,7 +8,7 @@ export { UpdatePlanItemScreen } from './updatePlanItem/UpdatePlanItemScreen';
 export { DashboardScreen } from './dashboard/DashboardScreen';
 export { UpdatePlanScreen } from './updatePlan/UpdatePlanScreen';
 export { StudentSettingsScreen } from './studentSettings/StudentSettingsScreen';
-export { RunPlanListScreen } from './runPlan/runPlanList/RunPlanListScreen';
-export { RunSubPlanListScreen } from './runPlan/runPlanList/RunSubPlanListScreen';
-export { RunPlanSlideScreen } from './runPlan/runPlanSlide/RunPlanSlideScreen';
+export { RunPlanListScreen } from './runPlan/ListMode/RunPlanListScreen';
+export { RunSubPlanListScreen } from './runPlan/ListMode/RunSubPlanListScreen';
+export { RunPlanSlideScreen } from './runPlan/SlideMode/RunPlanSlideScreen';
 export { UpdatePlanSubItemScreen } from './updatePlanItem/UpdatePlanSubItemScreen';
